Create CONTRIBUTING.md (#101)
* Create CONTRIBUTING.md

I copied this from AutomatedLab/AutomatedLab because I really liked it and thought it might be a good place to start a conversation. I made some minor changes to fit Polaris but feel free to tweak or throw away the whole thing.

I just think it would be nice to have some guidelines set down since I'm in here making a lot of changes lately and hope others will jump in too! :)

* Update CONTRIBUTING.md