Better check for timeoutManager api (#125)
* Better check for timeoutManager api

The previous check was attempting to check if timeoutmanager is available on the object by determining the OS architecture. I assumed if you were Linux, you were using .net core which meant your version of HttpListener did not contain the timeoutmanager.

Flawed logic... Now I am just checking to see if the member exists and if it does we use it.

Should resolve issue #124

* Returning check for platformID

* Checking for correct property

* Simpler check

Thanks for the docker image!